pyproject.toml
[project]
name = "question-gen-rl"
version = "0.1.0"
description = "Online RL for technical question generation using grok-4-1-fast-non-reasoning"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27.0",
    "pandas>=2.0.0",
    "pyarrow>=14.0.0",
    "python-dotenv>=1.0.0",
    "xai-sdk>=0.1.0",
]

[project.optional-dependencies]
train = [
    "skyrl-train",
    "skyrl-gym",
    "torch>=2.0.0",
    "vllm>=0.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/recruiter"]
